+++ CHANGES.md
@@ -10,6 +10,7 @@ Date: unreleased
 - Add helper for parse the query params: `catacumba.core/parse-queryparams`.
 - Add support for interceptors.
 - Add support for auth facilities.
+- Add support for extensible body parsing.
 - Removed support for automatic searching of `catacumba.properties` file for basedir.
 - Start using boot-clj as replacement of leiningen (only for main library, examples
   stil using leiningen beacuse users are more familiar with it).